converted to a better file structure
updated file structure
