+++ web_app/db.py
@@ -1,6 +1,5 @@
-from flask import Flask
 from flask_pymongo import pymongo
 
-CONNECTION_STRING = "mongodb+srv://nikoswee:apple1234@rent-a-cam.tf5um47.mongodb.net/?retryWrites=true&w=majority"
-client = pymongo.MongoClient(CONNECTION_STRING)
+DB_URI = "mongodb+srv://nikoswee:apple1234@rent-a-cam.tf5um47.mongodb.net/?retryWrites=true&w=majority"
+client = pymongo.MongoClient(DB_URI, tls=True)
 db = client.get_database('Cam-A-Rent')
