+++ web_app/decorator.py
@@ -1,0 +1,12 @@
+from flask import session, redirect, url_for
+from functools import wraps
+
+
+def login_required(f):
+    @wraps(f)
+    def wrap(*args, **kwargs):
+        if 'logged_in' in session:
+            return f(*args, **kwargs)
+        else:
+            return redirect(url_for("homePage"))
+    return wrap
