+++ web_app/static/js/scripts.js
@@ -11,7 +11,7 @@ $("form[name=signup_form").submit(function (e) {
         data: data,
         dataType: "json",
         success: function (resp) {
-            window.location.href = "/dashboard/";
+            window.location.href = "/user/dashboard";
         },
         error: function (resp) {
             $error.text(resp.responseJSON.error).removeClass("error--hidden");
@@ -34,7 +34,7 @@ $("form[name=login_form").submit(function (e) {
         data: data,
         dataType: "json",
         success: function (resp) {
-            window.location.href = "/dashboard/";
+            window.location.href = "/user/dashboard";
         },
         error: function (resp) {
             $error.text(resp.responseJSON.error).removeClass("error--hidden");
@@ -44,24 +44,3 @@ $("form[name=login_form").submit(function (e) {
     e.preventDefault();
 });
 
-$("form[name=uploadlisting_form").submit(function (e) {
-
-    var $form = $(this);
-    var $error = $form.find(".error");
-    var data = $form.serialize();
-
-    $.ajax({
-        url: "/user/uploadListing",
-        type: "POST",
-        data: data,
-        dataType: "json",
-        success: function (resp) {
-            window.location.href = "/dashboard/";
-        },
-        error: function (resp) {
-            $error.text(resp.responseJSON.error).removeClass("error--hidden");
-        }
-    });
-
-    e.preventDefault();
-});
