+++ web_app/templates/allProductView.html
@@ -13,16 +13,14 @@
             </div>
         </div>
     </section>
-
-    <form action="/search" method="GET" class="d-flex mt-3">
-        <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search" name="searchText"
-            id="searchText">
-        <button class="btn btn-outline-success" type="submit">Search</button>
-    </form>
-
-
     <div class="album py-5 bg-light">
         <div class="container">
+                <form action="/product/search/" method="GET" class="d-flex mt-3">
+                    <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search" name="searchText"
+                        id="searchText">
+                    <button class="btn btn-success" type="submit">Search</button>
+                </form>
+            <p>
             <div class="row row-cols-1 row-cols-sm-2 row-cols-md-3 g-3">
                 {% for item in product %}
                 <div class="col">
