+++ web_app/templates/dashboard.html
@@ -7,7 +7,7 @@
             <p class="center">Welcome {{session['user']['username']}} !
             </p>
             <div class="center">
-                <a class="btn btn-primary" href="/user/signout" role="button">Sign Out</a>
+                <a class="btn btn-primary" href="/user/signout/" role="button">Sign Out</a>
             </div>
         </div>
 
@@ -20,8 +20,8 @@
         <div class="card">
             <h2 class="center">Listings</h2>
             <div class="center">
-                <br><a class="btn btn-primary" href="/listing" role="button">View Your Listings</a></br>
-                <br><a class="btn btn-primary" href="/uploadListing" role="button">Create Listing</a></br>
+                <br><a class="btn btn-primary" href="/user/viewListing/" role="button">View Your Listings</a></br>
+                <br><a class="btn btn-primary" href="/user/uploadListing/" role="button">Create Listing</a></br>
             </div>
         </div>
     </div>
