+++ web_app/templates/index.html
@@ -10,7 +10,7 @@
   <title>Cam-A-Rent</title>
   <link rel="canonical" href="https://getbootstrap.com/docs/5.2/examples/album/">
 
-  <link href="/static/css/bootstrap.min.css" rel="stylesheet"
+  <link href="{{ url_for('static',filename='css/bootstrap.min.css') }}" rel="stylesheet"
     integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
 
   <style>
@@ -77,7 +77,7 @@
       <div class="offcanvas offcanvas-end text-bg-dark" tabindex="-1" id="offcanvasNavbarDark"
         aria-labelledby="offcanvasNavbarDarkLabel">
         <div class="offcanvas-header">
-          <h5 class="offcanvas-title" id="offcanvasNavbarDarkLabel">Offcanvas</h5>
+          <h5 class="offcanvas-title" id="offcanvasNavbarDarkLabel">Cam-A-Rent</h5>
           <button type="button" class="btn-close btn-close-white" data-bs-dismiss="offcanvas"
             aria-label="Close"></button>
         </div>
@@ -94,27 +94,8 @@
               <a class="nav-link" href="/signup">Sign Up</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="/allProductView">Products</a>
+              <a class="nav-link" href="/product/overview/">Products</a>
             </li>
-            <li class="nav-item dropdown">
-              <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown"
-                aria-expanded="false">
-                Dropdown
-              </a>
-              <ul class="dropdown-menu">
-                <li><a class="dropdown-item" href="#">Action</a></li>
-                <li><a class="dropdown-item" href="#">Another action</a></li>
-                <li>
-                  <hr class="dropdown-divider">
-                </li>
-                <li><a class="dropdown-item" href="#">Something else here</a></li>
-              </ul>
-            </li>
-          </ul>
-          <form class="d-flex mt-3" role="search">
-            <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-            <button class="btn btn-outline-success" type="submit">Search</button>
-          </form>
         </div>
         {% elif session["logged_in"] == True %}
         <div class="offcanvas-body">
@@ -123,30 +104,11 @@
               <a class="nav-link active" aria-current="page" href="/">Home</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="/dashboard">Dashboard</a>
+              <a class="nav-link" href="/user/dashboard/">Dashboard</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="/allProductView">Products</a>
+              <a class="nav-link" href="/product/overview">Products</a>
             </li>
-            <li class="nav-item dropdown">
-              <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown"
-                aria-expanded="false">
-                Dropdown
-              </a>
-              <ul class="dropdown-menu">
-                <li><a class="dropdown-item" href="#">Action</a></li>
-                <li><a class="dropdown-item" href="#">Another action</a></li>
-                <li>
-                  <hr class="dropdown-divider">
-                </li>
-                <li><a class="dropdown-item" href="#">Something else here</a></li>
-              </ul>
-            </li>
-          </ul>
-          <form class="d-flex mt-3" role="search">
-            <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-            <button class="btn btn-outline-success" type="submit">Search</button>
-          </form>
         </div>
         {% endif %}
       </div>
@@ -158,9 +120,9 @@
   {%block content%}
   {# template #}
   {% endblock %}
-  <script src="/static/js/jquery.js"></script>
-  <script src="/static/js/scripts.js"></script>
-  <script src="/static/js/bootstrap.bundle.min.js"
+  <script src="{{ url_for('static',filename='js/jquery.js') }}"></script>
+  <script src="{{ url_for('static',filename='js/scripts.js') }}"></script>
+  <script src="{{ url_for('static',filename='js/bootstrap.bundle.min.js') }}"
     integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
     crossorigin="anonymous"></script>
 </body>
