+++ web_app/templates/listing.html
@@ -15,11 +15,11 @@
               <p class="card-text" style="text-align: center">${{item.stock}}/day</p>
               <div class="d-flex justify-content-between align-items-center">
                 <div class="btn-group">
-                  <form action="{{url_for('updateProduct', id=item._id)}}" method="POST">
+                  <form action="{{url_for('user_bp.delListing', id=item._id)}}" method="POST">
                     <button type="submit" class="btn btn-sm btn-outline-secondary" name="editObjID" id="editObjID"
-                    value="{{item._id}}">Delete</button>
+                    value="{{item._id}}">Edit</button>
                     <button type="submit" class="btn btn-sm btn-outline-secondary" name="delObjID" id="delObjID"
-                      value="{{item._id}}">Delete</button>
+                    value="{{item._id}}">Delete</button>
                   </form>
                 </div>
               </div>
