--- web_app/user/routes.py
@@ -1,18 +1,0 @@
-from flask import Flask
-from app import app
-from user.models import User
-
-
-@app.route('/user/signup/', methods=['POST'])
-def signup():
-    return User().signup()
-
-
-@app.route('/user/login/', methods=['POST'])
-def login():
-    return User().login()
-
-
-@app.route('/user/signout/', methods=['POST', 'GET'])
-def signout():
-    return User().signout()
